No issue: Add scripts/places_missing_data.
Given a DB, this will find places with missing data.

I intend to follow this up with additional commits that:

1. Provide a UI that will allow us to create our own crosswalk: it will pull
place from the list of places missing data, prompt the user to enter in
urls for TA, wikipedia, and the place's website, and store these on the
server.
2. Uses these new IDs to fetch provider data into the DB: I expect
this will either add to a pre-cached DB or interface directly with our
places fetching. Suggestions welcome on approaches!

Other considerations:
* Should we build an interface to overwrite place data with custom data? (e.g.
so antlam could set custom reviews or trivia facts). Firebase console doesn't
work now that we have so much data. :(
* How would these tools work if we wanted to customize the place list? (e.g.
entirely hand-crafted or we remove undesired places)

wrt ^, it could help me to get more context on how we currently fetch places.