Issue #87 - Recursively subdivide base Yelp results
This makes our result set much more complete. Because the number of
locations significantly increases with these changes, the crawl script
has been changed to perform only an initial crawl from yelp. We'll need
to create another crawl script that fetches from each of the providers
to build on the base results.