Issue #85: with 2to3 + edits, add unicode changes.
strip_accents will throw in unicodedata.normalize if not given a unicode string
so we no longer need to assert. We don't want to do the conversion ourself
because stripping accents from a byte string doesn't make sense (the
conversions depending on the encoding, e.g. utf-8).

Everything is a unicode string now so we don't need to do the conversions when
calling into the functions.