Issue #85: Rm default encoding hack.
This failed at runtime while executed crawl_base.

I don't know why it was here but http://stackoverflow.com/a/17628350/2219998
implies there may have been an encoding mismatch between the loading files -
maybe this was because we were running some python2/3 hybrid?

Also noted is that this hack could break code that relies on ascii being the
default, even third-party code - it's a good thing we're getting rid of this!