Closes #85: Correct readme for python3 support.
To verify this conversion correct, I've run:
- pytest
- python -m scripts.crawl_base

This is the main code we're using on the server at this point - I figure we can
clean up anything else on the server as we get to it.

Developers will need to create a new virtual env for python3 and maybe destroy
their current virtualenvs - running `pytest` will crash, and I think it's
because there are two conflicting virtualenvs and `pytest tests` must be run
instead.